SFI Fix For DMs

+++ sigma/plugins/moderation/sfiexec.py
@@ -2,15 +2,16 @@ from sigma.core.permission import check_admin
 
 
 async def sfiexec(ev, message, args):
-    if 'discord.gg' in message.content:
-        if check_admin(message.author, message.channel):
-            return
-        else:
-            active = ev.db.get_settings(message.server.id, 'BlockInvites')
-            if active:
-                try:
-                    await ev.bot.delete_message(message)
-                except Exception as e:
-                    ev.log.error(e)
-            else:
+    if message.server:
+        if 'discord.gg' in message.content:
+            if check_admin(message.author, message.channel):
                 return
+            else:
+                active = ev.db.get_settings(message.server.id, 'BlockInvites')
+                if active:
+                    try:
+                        await ev.bot.delete_message(message)
+                    except Exception as e:
+                        ev.log.error(e)
+                else:
+                    return
